+++ app/src/androidTest/java/com/aa/draftt/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.draftt
+package com.aa.draftt
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
