+++ app/src/main/AndroidManifest.xml
@@ -3,20 +3,20 @@
     package="com.example.draftt">
 
     <application
-        android:name=".App"
+        android:name="com.aa.draftt.App"
         android:allowBackup="true"
         android:icon="@mipmap/draftt_app_icon"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/draftt_app_icon_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Draftt">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.aa.draftt.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".auth.views.AuthActivity">
+        <activity android:name="com.aa.draftt.auth.views.AuthActivity">
         </activity>
     </application>
