+++ app/src/main/java/com/aa/draftt/App.kt
@@ -1,4 +1,4 @@
-package com.example.draftt
+package com.aa.draftt
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
