+++ app/src/main/java/com/aa/draftt/Utils.kt
@@ -1,4 +1,4 @@
-package com.example.draftt
+package com.aa.draftt
 
 import android.content.Context
 import android.view.View
