+++ app/src/main/java/com/aa/draftt/auth/repositories/FirebaseAuthRepository.kt
@@ -1,4 +1,4 @@
-package com.example.draftt.auth.repositories
+package com.aa.draftt.auth.repositories
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
