+++ app/src/main/java/com/aa/draftt/auth/viewmodels/LoginViewModel.kt
@@ -1,24 +1,24 @@
-package com.example.draftt.auth.viewmodels
+package com.aa.draftt.auth.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.draftt.auth.repositories.AuthResult
-import com.example.draftt.auth.repositories.FirebaseAuthRepository
+import com.aa.draftt.auth.repositories.AuthResult
+import com.aa.draftt.auth.repositories.FirebaseAuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class SignUpViewModel @Inject constructor(private val authRepository: FirebaseAuthRepository) :
+class LoginViewModel @Inject constructor(private val authRepository: FirebaseAuthRepository) :
     ViewModel() {
 
     val authResult: LiveData<AuthResult> = authRepository.authResult
 
-    fun signup(email: String, password: String) {
+    fun login(email: String, password: String) {
         viewModelScope.launch(Dispatchers.IO) {
-            authRepository.signup(email, password)
+            authRepository.login(email, password)
         }
     }
