+++ app/src/main/java/com/aa/draftt/auth/views/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.example.draftt.auth.views
+package com.aa.draftt.auth.views
 
 import android.content.Context
 import android.os.Bundle
@@ -12,12 +12,11 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import com.example.draftt.R
-import com.example.draftt.Utils
-import com.example.draftt.auth.viewmodels.LoginViewModel
+import com.aa.draftt.Utils
+import com.aa.draftt.auth.viewmodels.LoginViewModel
 import com.example.draftt.databinding.FragmentLoginBinding
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
